+++ packages/backup/interfaces.go
@@ -8,6 +8,7 @@ import (
 	v3 "github.com/cloudogu/k8s-blueprint-lib/v3/api/v3"
 	componentV1 "github.com/cloudogu/k8s-component-lib/api/v1"
 	batchv1 "k8s.io/api/batch/v1"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -41,4 +42,10 @@ type componentClient interface {
 
 type cronJobClient interface {
 	Get(ctx context.Context, name string, opts metav1.GetOptions) (*batchv1.CronJob, error)
+}
+
+type configmapClient interface {
+	// Get takes name of the component, and returns the corresponding component object, and an error if there is any.
+	Get(ctx context.Context, name string, opts metav1.GetOptions) (*v1.ConfigMap, error)
+	Update(ctx context.Context, config *v1.ConfigMap, opts metav1.UpdateOptions) (*v1.ConfigMap, error)
 }